Fixed some bugs with chan and readers
